deploy/requirements.txt
fastapi>=0.115
uvicorn>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
httpx>=0.27
boto3>=1.35
anthropic>=0.39
langgraph>=0.2
python-dotenv>=1.0
